ssr fix with category slider component
